% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.gf.rep.R
\name{def.gf.rep}
\alias{def.gf.rep}
\title{Gap-fill with replicate sensor}
\usage{
def.gf.rep(dataGf, meas = !base::is.na(dataGf[[NameVarGf]]), NameVarGf,
  NameVarRep, Wndw, NumSampMin = 5, RtioRngRepMax = NULL, LvlPred = 0.95)
}
\arguments{
\item{dataGf}{Numeric data frame including all variables needed for gap-filling (target sensor data and replicate sensor data).}

\item{meas}{Logical vector same length as nrow(dataGf), where TRUE indicates originally-measured values of the gap-filled variable. Defaults to TRUE for non-NA indices of the gap-filled variable in dataGf.}

\item{NameVarGf}{Character string. The name of the variable to gap-fill}

\item{NameVarRep}{Character vector. One or more strings indicating the names of the variables that represent replicate sensor measurements of the gap-filled variable.}

\item{Wndw}{Integer value. Window [+- number of points] around each missing value for computing the regression with replicate sensor data}

\item{NumSampMin}{Integer value. The minimum number of samples required to perform the regression with the replicate sensor. Defaults to 5.}

\item{RtioRngRepMax}{Numeric value [fraction]. The maximum fraction of the range of replicate sensor data (used in the regression) 
that the replicate sensor value at the target gap may exceed be in order to fill with the predicted value. For example, if
the replicate sensor data used in the regression ranged from 0-10, and RtioRngRepMax was 0.1, a predicted target sensor value 
would only be accepted if the replicate sensor value at the gap was no less than -1 and no greater than 11 (since 10% 
of 10 is 1). Can be negative (for a range narrower than the data range). If NULL, this condition is not applied. Defaults to NULL.}

\item{LvlPred}{Numeric value between 0 and 1 (non-inclusive). Tolerance/confidence level of the prediction (from which the prediction interval is calculated). Defaults to .95.}
}
\value{
A data frame with the variables:\cr
\code{pred} The gap-filled dataset
\code{lwr} The lower prediction interval for each gap-filled value. NA for non-filled values
\code{upr} The upper prediction interval for each gap-filled value, NA for non-filled values
\code{se} THe standard error of the prediction
\code{df} The degrees of freedom of the regression used to fill each prediction (n-2)
}
\description{
Gap-fill meteorological data with data from a replicate sensor, using a windowed regression approach
}
\examples{
Currently None
}
\references{
None
}
\author{
Cove Sturtevant
}
\keyword{Currently}
\keyword{None.}
